+++ src/cli/dcos_aws/commands/create.py
@@ -187,6 +187,7 @@ def create(
                     **cluster.base_config,
                     **extra_config,
                 },
+                ip_detect_path=cluster_backend.ip_detect_path,
             )
     except CalledProcessError as exc:
         click.echo('Error installing DC/OS.', err=True)