+++ CHANGELOG.rst
@@ -6,7 +6,7 @@ Changelog
 Next
 ----
 
-- Add ``ip_detect_path`` to the ``Cluster.from_nodes`` function.
+- Add ``ip_detect_path`` to the each ``ClusterBackend`` as a property and to each install DC/OS function as a parameter.
 
 2018.07.23.1
 ------------